requirements.txt
PySide6>=6.5
types-PySide6>=6.10
SQLAlchemy>=2.0
invoke>=2.2.1
pyyaml>=6.0.3
tomlkit>=0.13.3
pyinstaller>=6.17
pytest>=9.0
pytest-mock>=3.15.1
pytest-cov>=7.0.0
pytest-qt>=4.5.0
pillow>=12.0